requirements.txt
youtube-transcript-api
MonsterUI
python-fasthtml